+++ components/[pageId]/DocumentPage/components/BountyProperties.tsx
@@ -84,7 +84,6 @@ export default function BountyProperties (props: {children: ReactNode, readOnly?
   const assignedRoleSubmitters = permissions?.bountyPermissions?.submitter?.filter(p => p.group === 'role').map(p => p.id as string) ?? [];
   const selectedReviewerUsers = permissions?.bountyPermissions?.reviewer?.filter(p => p.group === 'user').map(p => p.id as string) ?? [];
   const selectedReviewerRoles = permissions?.bountyPermissions?.reviewer?.filter(p => p.group === 'role').map(p => p.id as string) ?? [];
-  const [loggingIn, setLoggingIn] = useState(false);
 
   const canEdit = user && !readOnly && ((bounty.createdBy === user.id && bounty.status !== 'suggestion') || (bounty.status === 'suggestion' && isAdmin) || isAdmin);
 
@@ -372,10 +371,10 @@ export default function BountyProperties (props: {children: ReactNode, readOnly?
       }}
     >
       {permissions && (user?.id === bounty.createdBy || isAdmin) && (
-      <BountyHeader
-        bounty={bounty}
-        permissions={permissions}
-      />
+        <BountyHeader
+          bounty={bounty}
+          permissions={permissions}
+        />
       )}
       <Stack flexDirection='row' justifyContent='space-between' gap={2} alignItems='center'>
         {
@@ -432,17 +431,17 @@ export default function BountyProperties (props: {children: ReactNode, readOnly?
       />
       {children}
       {permissions && bounty.status !== 'suggestion' && (
-      <>
-        <BountySubmissionsTable
-          bounty={currentBounty}
-          permissions={permissions}
-        />
-        <Divider
-          sx={{
-            my: 1
-          }}
-        />
-      </>
+        <>
+          <BountySubmissionsTable
+            bounty={currentBounty}
+            permissions={permissions}
+          />
+          <Divider
+            sx={{
+              my: 1
+            }}
+          />
+        </>
       )}
 
       {permissions?.userPermissions?.review && bounty.status === 'suggestion' && bounty.createdBy !== user?.id && (
