+++ components/bounties/components/BountyApplicantForm/index.tsx
@@ -1,0 +1,3 @@
+import Form from './BountyApplicantForm';
+
+export default Form;
